+++ src/main/java/com/saral/reporting/model/RoleAssignment.java
@@ -14,24 +14,24 @@ public class RoleAssignment {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "user_id")
-	private Integer userId;
+	private Long userId;
 
 	@Column(name = "roleId" , nullable = false)
-	private Integer roleId; // integer NOT NULL,
+	private Long roleId; // integer NOT NULL,
 	@Column(name = "user_role_id",  nullable = false)
 	private Long useRoleId; // bigint NOT NULL,
 	@Column(name = "user_loc_id")
 	private Long userLocId; // bigint,
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
-	public Integer getRoleId() {
+	public Long getRoleId() {
 		return roleId;
 	}
-	public void setRoleId(Integer roleId) {
+	public void setRoleId(Long roleId) {
 		this.roleId = roleId;
 	}
 	public Long getUseRoleId() {
